# Refs #32290 -- Optimized construct_relative_path() by delay computing has_quotes.