Small little thing I noted while checking Collin's changes in #15140; `has_quotes` is unused until the very end of `construct_relative_path` (used by `do_extends`, `do_include` and `IncludeNode`), at which point it is known to be a quoted relative path.

For the common case of using "absolute" (relative to the template loader root...) paths, this seems unnecessary, and a small gain can be had by pushing it further down (and removing the `startswith` calls)

Before:
```
In [2]: %timeit construct_relative_path('dir1/two.html', 'dir2/two.html')
445 ns ± 6.97 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [3]: %timeit construct_relative_path('dir1/two.html', '../two.html')
4.1 µs ± 35.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

After:
```
In [2]: %timeit construct_relative_path('dir1/two.html', 'dir2/two.html')
263 ns ± 6.16 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [3]: %timeit construct_relative_path('dir1/two.html', '../two.html')
4.09 µs ± 50.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

I did also toy with checking for quotes before doing the `.strip()` call, but it came out as a wash; removing two characters from the start/end of a relatively short string seems thankfully well optimised vs. manually peeking.

-----

I've not created a ticket for this as it's basically moving a variable's declaration, but let me know if you'd like one. Naturally it can also just be dismissed without one if you'd prefer.